Added separate presenter for comments
