+++ src/model/comments-model.js
@@ -1,5 +1,6 @@
 import Observable from '../framework/observable.js';
 import { ErrorType, UpdateType } from '../const.js';
+import { snakeCaseKeysToCamelCase } from '../utils/common.js';
 
 export default class CommentsModel extends Observable {
   #moviesApiService = null;
@@ -22,27 +23,34 @@ export default class CommentsModel extends Observable {
     } catch(err) {
       this.#comments = ErrorType.COMMENTS_ERROR;
     }
-
     this._notify(UpdateType.INIT, this.#movie);
   };
 
-  addComment = (updateType, newComment) => {
-    this.#movie.comments = [
-      newComment.id,
-      ...this.#movie.comments,
-    ];
-
-    this._notify(updateType, this.#movie);
+  addComment = async (updateType, comment) => {
+    try {
+      const response = await this.#moviesApiService.addComment(this.#movie, comment);
+      const newComment = this.#adaptToClient(response);
+      this.#comments = [newComment, ...this.#comments];
+      this._notify(updateType, this.#movie);
+    } catch(err) {
+      throw new Error('Can\'t add comment');
+    }
   };
 
-  deleteComment = (updateType, commentID) => {
+  deleteComment = async (updateType, commentID) => {
     const index = this.#movie.comments.findIndex((comment) => comment === commentID);
 
-    this.#movie.comments = [
-      ...this.#movie.comments.slice(0, index),
-      ...this.#movie.comments.slice(index + 1),
-    ];
-
-    this._notify(updateType, this.#movie);
+    try {
+      await this.#moviesApiService.deleteComment(commentID);
+      this.#comments = [
+        ...this.#comments.slice(0, index),
+        ...this.#comments.slice(index + 1),
+      ];
+      this._notify(updateType, this.#movie);
+    } catch(err) {
+      throw new Error('Can\'t delete comment');
+    }
   };
+
+  #adaptToClient = (comment) => snakeCaseKeysToCamelCase(comment);
 }
